requirements.txt
absl-py==2.3.0
aiohappyeyeballs==2.6.1
aiohttp==3.12.7
aiosignal==1.3.2
annotated-types==0.7.0
anyascii==0.3.2
attrs==25.3.0
audioread==3.0.1
babel==2.17.0
bangla==0.0.5
beautifulsoup4==4.13.4
blinker==1.9.0
blis==1.2.1
bnnumerizer==0.0.2
bnunicodenormalizer==0.1.7
Brotli==1.1.0
catalogue==2.0.10
certifi==2025.4.26
cffi==1.17.1
charset-normalizer==3.4.2
click==8.2.1
cloudpathlib==0.21.1
confection==0.1.5
contourpy==1.3.2
coqpit==0.0.17
cryptography==45.0.3
cycler==0.12.1
cymem==2.0.11
Cython==3.1.1
dateparser==1.1.8
decorator==5.2.1
docopt==0.6.2
einops==0.8.1
encodec==0.1.1
es_core_news_md @ https://github.com/explosion/spacy-models/releases/download/es_core_news_md-3.8.0/es_core_news_md-3.8.0-py3-none-any.whl#sha256=478b8bb3f3e8eb149192f7d80e7cd64f990b7e9bccbc5329df41a57e86326be2
filelock==3.18.0
Flask==3.1.1
fonttools==4.58.1
frozenlist==1.6.0
fsspec==2025.5.1
g2pkk==0.1.2
grpcio==1.72.1
gruut==2.2.3
gruut-ipa==0.13.0
gruut_lang_de==2.0.1
gruut_lang_en==2.0.1
gruut_lang_es==2.0.1
gruut_lang_fr==2.0.2
h11==0.16.0
h2==4.2.0
hangul-romanize==0.1.0
hf-xet==1.1.2
hpack==4.1.0
huggingface-hub==0.32.4
hyperframe==6.1.0
idna==3.10
inflect==7.5.0
itsdangerous==2.2.0
jamo==0.4.1
jieba==0.42.1
Jinja2==3.1.6
joblib==1.5.1
jsonlines==1.2.0
kaitaistruct==0.10
kiwisolver==1.4.8
langcodes==3.5.0
language_data==1.3.0
lazy_loader==0.4
librosa==0.11.0
llvmlite==0.44.0
marisa-trie==1.2.1
Markdown==3.8
markdown-it-py==3.0.0
MarkupSafe==3.0.2
matplotlib==3.10.3
mdurl==0.1.2
more-itertools==10.7.0
mpmath==1.3.0
msgpack==1.1.0
multidict==6.4.4
murmurhash==1.0.13
networkx==2.8.8
nltk==3.9.1
num2words==0.5.14
numba==0.61.2
numpy==1.26.4
openai-whisper @ git+https://github.com/openai/whisper.git@dd985ac4b90cafeef8712f2998d62c59c3e62d22
outcome==1.3.0.post0
packaging==25.0
pandas==1.5.3
pillow==11.2.1
platformdirs==4.3.8
pooch==1.8.2
preshed==3.0.10
propcache==0.3.1
protobuf==6.31.1
psutil==7.0.0
pyasn1==0.6.1
pycparser==2.22
pydantic==2.11.5
pydantic_core==2.33.2
Pygments==2.19.1
pynndescent==0.5.13
pyOpenSSL==25.1.0
pyparsing==3.2.3
pypinyin==0.54.0
pysbd==0.3.4
PySocks==1.7.1
python-crfsuite==0.9.11
python-dateutil==2.9.0.post0
pytz==2025.2
PyYAML==6.0.2
regex==2024.11.6
requests==2.32.3
rich==14.0.0
safetensors==0.5.3
scikit-learn==1.6.1
scipy==1.15.3
selenium==4.33.0
selenium-wire==5.1.0
sentencepiece==0.2.0
shellingham==1.5.4
six==1.17.0
smart-open==7.1.0
sniffio==1.3.1
sortedcontainers==2.4.0
soundfile==0.13.1
soupsieve==2.7
soxr==0.5.0.post1
spacy==3.8.7
spacy-legacy==3.0.12
spacy-loggers==1.0.5
srsly==2.5.1
SudachiDict-core==20250515
SudachiPy==0.6.10
sympy==1.14.0
tensorboard==2.19.0
tensorboard-data-server==0.7.2
thinc==8.3.4
threadpoolctl==3.6.0
tiktoken==0.9.0
tokenizers==0.21.1
torch==2.7.0
torchaudio==2.7.0
tqdm==4.67.1
trainer==0.0.36
transformers==4.52.4
trio==0.30.0
trio-websocket==0.12.2
TTS==0.22.0
typeguard==4.4.2
typer==0.16.0
typing-inspection==0.4.1
typing_extensions==4.13.2
tzlocal==5.3.1
umap-learn==0.5.7
Unidecode==1.4.0
urllib3==2.4.0
wasabi==1.1.3
weasel==0.4.1
websocket-client==1.8.0
Werkzeug==3.1.3
wrapt==1.17.2
wsproto==1.2.0
yarl==1.20.0
zstandard==0.23.0
